Clarified data plane url variable
The data plane url parameter is confusing to me since it looks like it's meant to append a path to the hostname string, but it's not valid javascript and it's not totally clear what the intended modification here is. I think what you guys mean to show is that the path "/v1/batch" should be appended to the dataplane hostname URL, but I think it would be clearer to show that as a string template.